google_calendar/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
python-multipart==0.0.5
google-auth==2.6.6
google-auth-oauthlib==0.5.1
google-api-python-client==2.44.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib